+++ CHANGELOG.md
@@ -5,6 +5,13 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Changed struct `Configuration` to enum `Configs` and defined configuration by
+  module.
+
 ## [0.9.3] - 2024-02-16
 
 ### Added
@@ -248,6 +255,7 @@ without relying on the content of the response.
 
 - `send_frame` and `recv_frame` to send and receive length-delimited frames.
 
+[Unreleased]: https://github.com/petabi/oinq/compare/0.9.3...main
 [0.9.3]: https://github.com/petabi/oinq/compare/0.9.2...0.9.3
 [0.9.2]: https://github.com/petabi/oinq/compare/0.9.1...0.9.2
 [0.9.1]: https://github.com/petabi/oinq/compare/0.9.0...0.9.1